+++ src/pacasam/samplers/targetted.py
@@ -1,13 +1,31 @@
+import logging
+import warnings
 import geopandas as gpd
 import pandas as pd
 from typing import Dict
-from pacasam.connectors.connector import FILE_ID_COLNAME
+from pacasam.connectors.connector import FILE_ID_COLNAME, Connector, PATCH_ID_COLNAME
 from pacasam.samplers.algos import sample_with_stratification
-from pacasam.samplers.sampler import Sampler
+from pacasam.samplers.sampler import Sampler, SPLIT_COLNAME
+from pacasam.samplers.spatial import SpatialSampler
+from math import floor
 
 
 class TargettedSampler(Sampler):
-    """A sampling to meet constraints - target prevalence of descriptor - sequentially."""
+    """A sampling to meet constraints - target prevalence of descriptor - sequentially.
+
+    Sampling is completed with spatial sampling if not enough patches are found.
+
+    """
+
+    def __init__(
+        self,
+        connector: Connector,
+        sampling_config: Dict,
+        log: logging.Logger = logging.getLogger(),
+        complete_with_spatial_sampling: bool = True,
+    ):
+        self.complete_with_spatial_sampling = complete_with_spatial_sampling
+        super().__init__(connector, sampling_config, log)
 
     def get_patches(self) -> gpd.GeoDataFrame:
         selection = []
@@ -17,12 +35,27 @@ class TargettedSampler(Sampler):
             patches = self._get_matching_patches(descriptor_name, descriptor_objectives)
             selection += [patches]
         selection = pd.concat(selection, ignore_index=True)
-        self.log.info(f"{self.name}: N={len(selection)} patches.")
+        selection = self.drop_duplicates_by_id_and_log_sampling_attrition(selection)
+        self.log.info(f"{self.name}: N={len(selection)} distinct patches selected to match TargettedSampler requirements.")
+
         if len(selection) > self.cf["target_total_num_patches"]:
-            self.log.warning(
+            warnings.warn(
                 f"Selected more than the desired total of N={self.cf['target_total_num_patches']}."
-                "If this is not desired, please reconsider your targets."
+                "If this is not desired, please reconsider your targets.",
             )
+        elif self.complete_with_spatial_sampling:
+            # Complete with spatial sampling and achieve the desired fraction of validation set
+            num_patches_to_add = self.cf["target_total_num_patches"] - len(selection)
+            final_num_patches_in_validation = floor(self.cf["frac_validation_set"] * self.cf["target_total_num_patches"])
+            num_patches_to_add_in_validation = final_num_patches_in_validation - len(selection[selection[SPLIT_COLNAME] == "val"])
+            self.cf["frac_validation_set"] = num_patches_to_add_in_validation / num_patches_to_add
+
+            # Complete with spatial sampling
+            ss = SpatialSampler(connector=self.connector, sampling_config=self.cf, log=self.log)
+            completion = ss.get_patches(num_to_sample=num_patches_to_add, current_selection_ids=selection[PATCH_ID_COLNAME])
+            selection = pd.concat([selection, completion])
+            self.log.info(f"{self.name}: completed targetted sampling with N={num_patches_to_add} additional patches.")
+
         return selection
 
     def _get_matching_patches(self, descriptor_name: str, descriptor_objectives: Dict):