+++ src/pacasam/samplers/triple.py
@@ -1,3 +1,4 @@
+import warnings
 import pandas as pd
 from pacasam.connectors.connector import PATCH_ID_COLNAME
 from pacasam.samplers.sampler import Sampler
@@ -10,14 +11,14 @@ class TripleSampler(Sampler):
     """Succession of Targetted, Diversity, and Completion sampling."""
 
     def get_patches(self) -> pd.Series:
-        ts = TargettedSampler(connector=self.connector, sampling_config=self.cf, log=self.log)
+        ts = TargettedSampler(connector=self.connector, sampling_config=self.cf, log=self.log, complete_with_spatial_sampling=False)
         targetted = ts.get_patches()
         targetted = ts.drop_duplicates_by_id_and_log_sampling_attrition(targetted)
 
         # Perform diversity sampling based on class histograms
         num_to_sample = (self.cf["target_total_num_patches"] - len(targetted)) // 2  # half of remaining patches
         if num_to_sample < 0:
-            self.log.warning(
+            warnings.warns(
                 f"Target dataset size of n={self.cf['target_total_num_patches']} patches achieved via targetted sampling single-handedly."
                 "\n This means the SUM OF CONSTRAINTS IS ABOVE 100%. Consider reducing constraints, and having a bigger dataset."
             )