+++ src/pacasam/utils.py
@@ -21,9 +21,9 @@ LOGGING_FORMATTER = logging.Formatter(fmt="%(asctime)s %(levelname)-8s %(message
 
 
 def setup_custom_logger():
-    log = logging.getLogger(__name__)
-    log.setLevel(logging.DEBUG)
-
+    logging.captureWarnings(True)
+    log = logging.getLogger()
+    log.setLevel(logging.INFO)
     screen_handler = logging.StreamHandler(stream=sys.stdout)
     screen_handler.setFormatter(LOGGING_FORMATTER)
     log.addHandler(screen_handler)